requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.9.1
aiosignal==1.3.2
attrs==25.1.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
docx==0.2.4
docx2pdf==0.1.8
easyocr==1.7.2
et_xmlfile==2.0.0
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.7.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.1
lazy_loader==0.4
Levenshtein==0.27.1
lxml==5.3.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.5
ninja==1.11.1.4
numpy==2.2.1
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pdf2image==1.17.0
pillow==11.2.1
propcache==0.2.1
pyarrow==21.0.0
pyclipper==1.3.0.post6
PyJWT==2.10.1
pytesseract==0.3.13
python-barcode==0.15.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-Levenshtein==0.27.1
pytz==2024.2
pywin32==310
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
soupsieve==2.6
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.6.11
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
twilio==9.1.1
typing_extensions==4.13.2
tzdata==2024.2
urllib3==2.3.0
yarl==1.18.3
